+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-18 18:34:18
-// Commit: chore(network): update dependencies and cleanup unused imports in network
+// Updated at: 2026-08-19 11:08:14
+// Commit: test(network): add unit tests for network business logic
 
 class networkComponentState {
-  final String commitRef = '3711';
-  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
-  final DateTime timestamp = DateTime.parse('2026-08-18T18:34:18+05:00');
+  final String commitRef = '3735';
+  final String lastMessage = 'test(network): add unit tests for network business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-19T11:08:14+05:00');
 }